Minor changes to default displays on the Weather page, cleaned some code. Deleted old Meteorologist file and renamed MeteorologsitStatic to take its place.
