+++ Weather.xaml
@@ -92,7 +92,7 @@
                 </Frame>
 
                 <Frame Grid.Row="1" Padding="10" CornerRadius="0" HasShadow="True" BorderColor="LightBlue">
-                    <Label x:Name="MetarLabel" Text="KATW 211605Z 2116/2212 07003KT P6SM -RA SCT022 OVC050 TEMPO 2116/2118 BKN004"/>
+                        <Label x:Name="MetarLabel" Text="Please enter an ICAO airport ID in the search bar and click the magnifying glass."/>
                 </Frame>
             </Grid>
 
@@ -102,10 +102,7 @@
                 </Frame>
 
                 <Frame Grid.Row="1" Padding="10" CornerRadius="0" HasShadow="True" BorderColor="LightBlue">
-                    <Label x:Name="TafLabel" Grid.Row="1" Text="FM211800 01009KT 5SM -RA BR OVC016
-                                              FM212200 35010G18KT P6SM -SHRA BKN020
-                                              FM220100 35008KT P6SM SCT030
-                                              FM220600 34007KT P6SM SCT035="/>
+                        <Label x:Name="TafLabel" Grid.Row="1" Text="Please enter an ICAO airport ID in the search bar and click the magnifying glass."/>
                 </Frame>
             </Grid>
         </Grid>
